Fix act title display
